Added external links to each item.
Additional copy to be displayed on the site above the listing of records and below the Label: The information presented in the table below is arranged alphabetically by title. Additional information is available on the <a href="[link]">[label]</a> website.

Note the University, Presidential, and Provostial Professorships does not have an external site to link to.